Organize HTML classes
