+++ app/src/jsMain/kotlin/diy/lingerie/dynamic_html/DynamicHtmlElement.kt
@@ -1,0 +1,11 @@
+package diy.lingerie.dynamic_html
+
+import org.w3c.dom.Element
+import org.w3c.dom.Node
+
+abstract class DynamicHtmlElement : DynamicHtmlNode() {
+    final override val rawNode: Node
+        get() = rawElement
+
+    abstract val rawElement: Element
+}
