+++ app/src/jsMain/kotlin/diy/lingerie/dynamic_html/DynamicHtmlNode.kt
@@ -1,0 +1,7 @@
+package diy.lingerie.dynamic_html
+
+import org.w3c.dom.Node
+
+abstract class DynamicHtmlNode {
+    abstract val rawNode: Node
+}
