+++ app/src/jsMain/kotlin/diy/lingerie/dynamic_html/DynamicWrapperElement.kt
@@ -1,0 +1,7 @@
+package diy.lingerie.dynamic_html
+
+import org.w3c.dom.Element
+
+class DynamicWrapperElement(
+    override val rawElement: Element,
+) : DynamicHtmlElement()
