+++ app/src/jsMain/kotlin/diy/lingerie/html/HtmlEvent.kt
@@ -1,0 +1,9 @@
+package diy.lingerie.html
+
+import org.w3c.dom.events.Event
+
+sealed class HtmlEvent {
+    interface Wrapper<E : HtmlEvent> {
+        fun wrap(rawEvent: Event): E
+    }
+}
