+++ app/src/jsMain/kotlin/diy/lingerie/html/collections/DomList.kt
@@ -1,0 +1,22 @@
+package diy.lingerie.html.collections
+
+interface DomList<out E> : BasicList<E> {
+    override val size: Int
+
+    override fun isEmpty(): Boolean {
+        return !isNotEmpty()
+    }
+
+    fun isNotEmpty(): Boolean {
+        return size > 0
+    }
+
+    override fun get(index: Int): E = getOrNull(index = index) ?: throw IndexOutOfBoundsException(
+        "Index $index is out of bounds for size $size",
+    )
+
+    fun getOrNull(index: Int): E?
+
+    val firstElement: E?
+        get() = getOrNull(0)
+}
