+++ app/src/jsMain/kotlin/diy/lingerie/html/collections/MutableDomList.kt
@@ -1,0 +1,3 @@
+package diy.lingerie.html.collections
+
+interface MutableDomList<E> : BasicMutableList<E>, DomList<E>
